Move project creation bits to a separate utility.
* removes IDEA-only classes from the small IDE project creation flow
